requirements.txt
mysqlclient
Django
django-registration-redux
django-debug-toolbar
django-recaptcha2
django-bootstrap3
django-countries
Pillow!=5.1.0
django-rules-light
django-mptt
django-pagination-bootstrap
django-markdownx
django-siteprefs
django-memoize